pyproject.toml
[project]
name = "mcp-explore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "mcp[cli]>=1.9.0",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "ruff>=0.11.10",
]
